requirements.txt
supervisely==6.69.0
numpy>=1.19.4
Pillow>=8.0.1
